Example route added
