+++ src/index.ts
@@ -10,23 +10,7 @@ import {
 } from './secrets';
 import routeBook from './routes/book';
 import routeUser from './routes/user';
-
-const EXAMPLE_BOOK: Book = {
-    id: 1,
-    library_user: 'John Doe',
-    title: 'Python Machine Learning: Machine Learning and Deep Learning with Python, scikit-learn, and TensorFlow 2, 3rd Edition',
-    author: 'Klabnik, Steve',
-    topic: 'Machine Learning',
-    isbn: '9781492032649',
-    location: 'Helsinki',
-};
-const EXAMPLE_BOOK_2: Book = { ...EXAMPLE_BOOK, id: 2, title: 'Book 2' };
-const EXAMPLE_BOOK_3: Book = { ...EXAMPLE_BOOK, id: 3, title: 'Book 3' };
-const EXAMPLE_BOOKS: Array<Book> = [
-    EXAMPLE_BOOK,
-    EXAMPLE_BOOK_2,
-    EXAMPLE_BOOK_3,
-];
+import routeExample from './routes/example';
 
 const app: Express = express();
 app.use(cors());
@@ -40,5 +24,6 @@ const pool = mysql.createPool({
 
 routeBook(app, pool);
 routeUser(app, pool);
+routeExample(app);
 
 export default app;
